add new api route for setting a new password after a reset
